+++ packages/ckeditor5-dev-webpack-plugin/lib/servetranslations.js
@@ -10,8 +10,6 @@ const rimraf = require( 'rimraf' );
 const fs = require( 'fs' );
 const path = require( 'path' );
 const semver = require( 'semver' );
-const { NormalModule } = require( 'webpack' );
-const { version: webpackVersion } = require( 'webpack/package.json' );
 const { RawSource, ConcatSource } = require( 'webpack-sources' );
 
 /**
@@ -94,11 +92,13 @@ module.exports = function serveTranslations( compiler, options, translationServi
 
 	// Load translation files and add a loader if the package match requirements.
 	compiler.hooks.compilation.tap( 'CKEditor5Plugin', compilation => {
-		getCompilationHooks( compilation ).tap( 'CKEditor5Plugin', ( context, module ) => {
+		getCompilationHooks( compiler, compilation ).tap( 'CKEditor5Plugin', ( context, module ) => {
 			const relativePathToResource = path.relative( cwd, module.resource );
 
 			if ( relativePathToResource.match( options.sourceFilesPattern ) ) {
-				module.loaders.push( {
+				// The `TranslateSource` loader must be added as the last one in the loader's chain,
+				// after any potential TypeScript file has already been compiled.
+				module.loaders.unshift( {
 					loader: path.join( __dirname, 'translatesourceloader.js' ),
 					options: { translateSource }
 				} );
@@ -196,15 +196,20 @@ function getPathToPackage( cwd, resource, packageNamePattern ) {
 /**
  * Returns an object with the compilation hooks depending on the Webpack version.
  *
- * @param {Object} compilation
+ * @param {webpack.Compiler} compiler
+ * @param {webpack.Compilation} compilation
  * @returns {Object}
  */
-function getCompilationHooks( compilation ) {
-	if ( semver.major( webpackVersion ) === 4 ) {
+function getCompilationHooks( compiler, compilation ) {
+	const { webpack } = compiler;
+
+	if ( semver.major( webpack.version ) === 4 ) {
 		return compilation.hooks.normalModuleLoader;
 	}
 
-	return NormalModule.getCompilationHooks( compilation ).loader;
+	// Do not import the `NormalModule` class directly. Find it in the current instance of webpack process.
+	// See: https://github.com/ckeditor/ckeditor5/issues/12887.
+	return webpack.NormalModule.getCompilationHooks( compilation ).loader;
 }
 
 /**